avcodec/ffv1enc: Remove 65536 pixel per slice limit for remap
About 1% better compression with large slices

+++ libavcodec/ffv1.h
@@ -108,14 +108,14 @@ typedef struct FFV1SliceContext {
     };
     int remap_count[4];
 
-    uint16_t   *bitmap  [4]; //float encode
+    uint32_t   *bitmap  [4]; //float encode
     uint16_t   *fltmap  [4]; //halffloat encode & decode
     uint32_t   *fltmap32[4]; //float decode
     unsigned int fltmap_size[4];
     unsigned int fltmap32_size[4];
     struct Unit {
         uint32_t val; //this is unneeded if you accept a dereference on each access
-        uint16_t ndx;
+        uint32_t ndx;
     } *unit[4];
 } FFV1SliceContext;
